client: refactor: reorganize services according to FSD architecture
