+++ client/src/app/entities/word/index.ts
@@ -1,2 +1,3 @@
 export type { Word, WordTranslation } from './model/word.types';
 export { WordApi } from '@entities/word/api/word.api';
+export { WordService } from '@entities/word/word.service';
