+++ client/src/app/shared/api/index.ts
@@ -1,3 +1,8 @@
-export { AuthApiService } from './auth.api.service';
-export { HttpApiService } from './http-api.service';
-export type { IUserLoginDTO, IUserLoginResponse, IUserRegisterDTO, IUserRegisterResponse } from './types/auth.dto';
+export { AuthApiService } from './auth/auth.api.service';
+export type {
+  IUserLoginDTO,
+  IUserLoginResponse,
+  IUserRegisterDTO,
+  IUserRegisterResponse,
+  IUserTokenPayload,
+} from './types/auth.dto';
