+++ client/src/app/shared/api/types/word-set.dto.ts
