+++ client/src/app/shared/infrastructure/http-api.service.ts
