+++ client/src/app/shared/infrastructure/index.ts
@@ -1,0 +1,4 @@
+export { TokenService } from './auth/token.service';
+export { HttpApiService } from './http-api.service';
+export { NetworkService } from './network.service';
+export { BreakpointService } from './ui/breakpoint.service';
