+++ client/src/app/shared/services/index.ts
@@ -1,1 +1,1 @@
-export { TokenService } from './token.service';
+// export { TokenService } from './token.service';
